+++ xrlint/rule.py
@@ -53,6 +53,21 @@ class RuleContext(ABC):
         """
 
 
+class RuleExit(Exception):
+    """The `RuleExit` is an exception that can be raised to
+    immediately cancel dataset node validation with the current rule.
+
+    Raise it from any of your `RuleOp` method implementations if further
+    node traversal doesn't make sense. Typical usage:
+
+    ```python
+    if something_is_not_ok:
+        ctx.report("Something is not ok.")
+        raise RuleExit
+    ```
+    """
+
+
 class RuleOp(ABC):
     """Define the specific rule verification operation."""
 
@@ -62,6 +77,8 @@ class RuleOp(ABC):
         Args:
             context: The current rule context.
             node: The dataset node.
+        Raises:
+            RuleExit: to exit rule logic and further node traversal
         """
 
     def data_array(self, context: RuleContext, node: DataArrayNode) -> None:
@@ -70,6 +87,8 @@ class RuleOp(ABC):
         Args:
             context: The current rule context.
             node: The data array (variable) node.
+        Raises:
+            RuleExit: to exit rule logic and further node traversal
         """
 
     def attrs(self, context: RuleContext, node: AttrsNode) -> None:
@@ -78,6 +97,8 @@ class RuleOp(ABC):
         Args:
             context: The current rule context.
             node: The attributes node.
+        Raises:
+            RuleExit: to exit rule logic and further node traversal
         """
 
     def attr(self, context: RuleContext, node: AttrNode) -> None:
@@ -86,6 +107,8 @@ class RuleOp(ABC):
         Args:
             context: The current rule context.
             node: The attribute node.
+        Raises:
+            RuleExit: to exit rule logic and further node traversal
         """