+++ CHANGES.md
@@ -5,6 +5,8 @@
   - enhanced "simple" output format by colors and links 
   - new xcube rule "increasing-time"
   - new xcube rule "data-var-colors"
+  - new `RuleExit` exception to exit rule logic and 
+    stop further node traversal
 
 - Version 0.0.2 (06.01.2025) 
   - more rules